requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
clarifai-grpc==8.0.0
dnspython==2.2.1
emoji==1.7.0
ephem==4.1.3
flake8==4.0.1
googleapis-common-protos==1.56.2
grpcio==1.46.3
idna==3.3
mccabe==0.6.1
protobuf==3.20.1
pycodestyle==2.8.0
pyflakes==2.4.0
pymongo==4.1.1
PySocks==1.7.1
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
